+++ src/widget/select-desktop/selectpicker.js
@@ -27,6 +27,8 @@ import event from '../../js/event';
 import { t } from 'enketo/translator';
 import '../../js/dropdown.jquery';
 
+const range = document.createRange();
+
 /**
  * Bootstrap Select picker that supports single and multiple selects
  * A port of https://github.com/silviomoreto/bootstrap-select
@@ -56,10 +58,11 @@ class DesktopSelectpicker extends Widget {
     }
 
     _init() {
-        const $select = $( this.element );
-        $select.css( 'display', 'none' );
-        const $template = this._createLi( this._getTemplate() );
-        this.$picker = $template.insertAfter( $select );
+        const select =  this.element;
+        select.style.display = 'none';
+        const template = this._getTemplate();
+        select.after( template );
+        this.picker = this.question.querySelector( '.bootstrap-select' );
         if ( this.props.readonly ) {
             this.disable();
         }
@@ -68,41 +71,36 @@ class DesktopSelectpicker extends Widget {
     }
 
     /**
-     * @return {string} HTML string
+     * @return {Element} HTML fragment
      */
     _getTemplate() {
-        return `
+        const template = range.createContextualFragment( `
         <div class="btn-group bootstrap-select widget clearfix">
             <button type="button" class="btn btn-default dropdown-toggle clearfix" data-toggle="dropdown">
-                <span class="selected">__SELECTED_OPTIONS</span><span class="caret"></span>
+                <span class="selected"></span><span class="caret"></span>
             </button>
-            <ul class="dropdown-menu" role="menu">__ADD_LI</ul>
-        </div>`;
+            <ul class="dropdown-menu" role="menu">${this._getLisHtml()}</ul>
+        </div>` );
+        this._showSelected( template.querySelector( '.selected' ) );
+
+        return template;
     }
 
     /**
-     * @param {string} template - The select template to use.
-     * @return {import('./type-def').jQuery} - The jQuery-wrapped template.
+     * Generates HTML text for <li> elements
      */
-    _createLi( template ) {
-        const li = [];
-        let liHtml = '';
+    _getLisHtml( ) {
         const inputAttr = this.props.multiple ? 'type="checkbox"' : `type="radio" name="${Math.random() * 100000}"`;
 
-        $( this.element ).find( 'option' ).each( function() {
-            li.push( {
-                label: $( this ).text(),
-                selected: $( this ).is( ':selected' ),
-                value: $( this ).attr( 'value' )
-            } );
-        } );
+        return [ ...this.element.querySelectorAll( 'option' ) ]
+            .map( option => {
+                const label = option.textContent;
+                const selected = option.matches( ':checked' );
+                const value = option.value;
+                if ( value ) {
+                    const checkedInputAttr = selected ? ' checked="checked"' : '';
+                    const checkedLiAttr = selected ? 'class="active"' : '';
 
-        if ( li.length > 0 ) {
-            template = template.replace( '__SELECTED_OPTIONS', this._createSelectedStr() );
-            for ( let i = 0; i < li.length; i++ ) {
-                if ( li[ i ].value ) {
-                    const checkedInputAttr = li[ i ].selected ? ' checked="checked"' : '';
-                    const checkedLiAttr = li[ i ].selected ? 'class="active"' : '';
                     /**
                      * e.g.:
                      * <li checked="checked">
@@ -113,47 +111,50 @@ class DesktopSelectpicker extends Widget {
                      *       </a>
                      *    </li>
                      */
-                    liHtml += `
-                    <li ${checkedLiAttr}>
-                        <a class="option-wrapper" tabindex="-1" href="#">
-                            <label>
-                                <input class="ignore" ${inputAttr}${checkedInputAttr} value="${li[ i ].value}" />
-                                <span class="option-label">${li[ i ].label}</span>
-                            </label>
-                        </a>
-                    </li>`;
+                    return `
+                        <li ${checkedLiAttr}>
+                            <a class="option-wrapper" tabindex="-1" href="#">
+                                <label>
+                                    <input class="ignore" ${inputAttr}${checkedInputAttr} value="${value}" />
+                                    <span class="option-label">${label}</span>
+                                </label>
+                            </a>
+                        </li>`;
+                } else {
+                    return '';
                 }
-            }
-        }
-
-        template = template.replace( '__ADD_LI', liHtml );
-
-        return $( template );
+            } ).join( '' );
     }
 
-
     /**
-     * Create text to show in closed picker
+     * Update text to show in closed picker
      *
-     * @return {string} - text to show in closed picker
+     * @param {Element} el - HTML element to show text in
      */
-    _createSelectedStr() {
-        const selectedLabels = [];
-        const $select = $( this.element );
-        $select.find( 'option:selected' ).each( function() {
-            if ( $( this ).attr( 'value' ).length > 0 ) {
-                selectedLabels.push( $( this ).text() );
-            }
-        } );
+    _showSelected( el ) {
+        const selectedLabels = [ ...this.element.querySelectorAll( 'option:checked' ) ]
+            .filter( option =>  option.getAttribute( 'value' ).length )
+            .map( option => option.textContent );
+
+        // keys for i18next parser to pick up:
+        // t( 'selectpicker.numberselected' );
 
         if ( selectedLabels.length === 0 ) {
-            return t( 'selectpicker.noneselected' );
+            // do not use variable for translation key to not confuse i18next-parser
+            el.textContent = t( 'selectpicker.noneselected' );
+            el.dataset.i18n =  'selectpicker.noneselected';
+            delete el.dataset.i18nNumber;
+
         } else if ( selectedLabels.length === 1 ) {
-            return selectedLabels[ 0 ];
+            el.textContent = selectedLabels[ 0 ];
+            delete el.dataset.i18n;
+            delete el.dataset.i18nNumber;
         } else {
-            return t( 'selectpicker.numberselected', {
-                number: selectedLabels.length
-            } );
+            const number = selectedLabels.length;
+            // do not use variable for translation key to not confuse i18next-parser
+            el.textContent = t( 'selectpicker.numberselected', { number } );
+            el.dataset.i18n = 'selectpicker.numberselected';
+            el.dataset.i18nNumber = number ;
         }
     }
 
@@ -163,7 +164,7 @@ class DesktopSelectpicker extends Widget {
     _clickListener() {
         const _this = this;
 
-        this.$picker
+        $( this.picker )
             .on( 'click', 'li:not(.disabled)', function( e ) {
                 const li = this;
                 const input = li.querySelector( 'input' );
@@ -179,9 +180,9 @@ class DesktopSelectpicker extends Widget {
                 }
 
                 if ( !_this.props.multiple ) {
-                    _this.$picker.find( 'li' ).removeClass( 'active' );
-                    getSiblingElementsAndSelf( option, 'option' ).forEach( el => { el.selected = false; } );
-                    _this.$picker.find( 'input' ).prop( 'checked', false );
+                    _this.picker.querySelectorAll( 'li' ).forEach( li=> li.classList.remove( 'active' ) );
+                    getSiblingElementsAndSelf( option, 'option' ).forEach( option => { option.selected = false; } );
+                    _this.picker.querySelectorAll( 'input' ).forEach( input  => input.checked = false );
                 } else {
                     //don't close dropdown for multiple select
                     e.stopPropagation();
@@ -205,7 +206,9 @@ class DesktopSelectpicker extends Widget {
                         input.checked = true;
                     }
 
-                    _this.$picker.find( '.selected' ).html( _this._createSelectedStr() );
+                    const showSelectedEl = _this.picker.querySelector( '.selected' );
+                    _this._showSelected( showSelectedEl );
+
                     select.dispatchEvent( new event.Change() );
                 }, 10 );
 
@@ -245,7 +248,7 @@ class DesktopSelectpicker extends Widget {
 
         // Focus on original element (form.goTo functionality)
         this.element.addEventListener( events.ApplyFocus().type, () => {
-            _this.$picker.find( '.dropdown-toggle' ).focus();
+            _this.picker.querySelector( '.dropdown-toggle' ).focus();
         } );
     }
 
@@ -253,7 +256,7 @@ class DesktopSelectpicker extends Widget {
      * Disables widget
      */
     disable() {
-        this.$picker[ 0 ].querySelectorAll( 'li' ).forEach( el => {
+        this.picker.querySelectorAll( 'li' ).forEach( el => {
             el.classList.add( 'disabled' );
             const input = el.querySelector( 'input' );
             // are both below necessary?
@@ -266,7 +269,7 @@ class DesktopSelectpicker extends Widget {
      * Enables widget
      */
     enable() {
-        this.$picker[ 0 ].querySelectorAll( 'li' ).forEach( el => {
+        this.picker.querySelectorAll( 'li' ).forEach( el => {
             el.classList.remove( 'disabled' );
             const input = el.querySelector( 'input' );
             input.disabled = false;
@@ -278,7 +281,7 @@ class DesktopSelectpicker extends Widget {
      * Updates widget
      */
     update() {
-        this.$picker.remove();
+        this.picker.remove();
         this._init();
     }
 }